updating

+++ DATABASE.sql
@@ -559,8 +559,8 @@ CREATE TABLE `player_depotitems` (
   `player_id` int(11) NOT NULL,
   `sid` int(11) NOT NULL COMMENT 'any given range eg 0-100 will be reserved for depot lockers and all > 100 will be then normal items inside depots',
   `pid` int(11) NOT NULL DEFAULT '0',
-  `itemtype` smallint(6) NOT NULL,
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `itemtype` int(11) NOT NULL,
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -587,8 +587,8 @@ CREATE TABLE `player_inboxitems` (
   `player_id` int(11) NOT NULL,
   `sid` int(11) NOT NULL,
   `pid` int(11) NOT NULL DEFAULT '0',
-  `itemtype` smallint(6) NOT NULL,
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `itemtype` int(11) NOT NULL,
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -602,8 +602,8 @@ CREATE TABLE `player_items` (
   `player_id` int(11) NOT NULL DEFAULT '0',
   `pid` int(11) NOT NULL DEFAULT '0',
   `sid` int(11) NOT NULL DEFAULT '0',
-  `itemtype` smallint(6) NOT NULL DEFAULT '0',
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `itemtype` int(11) NOT NULL DEFAULT '0',
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
 
@@ -686,8 +686,8 @@ CREATE TABLE `player_rewards` (
   `player_id` int(11) NOT NULL,
   `sid` int(11) NOT NULL,
   `pid` int(11) NOT NULL DEFAULT '0',
-  `itemtype` smallint(6) NOT NULL,
-  `count` smallint(5) NOT NULL DEFAULT '0',
+  `itemtype` int(11) NOT NULL,
+  `count` int(11) NOT NULL DEFAULT '0',
   `attributes` blob NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
